backend/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
future==0.18.2
gunicorn==22.0.0
lxml==6.0.2
numpy==2.0.2
opencv-python-headless
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pdf2image==1.17.0
pillow==11.3.0
pypdf==6.6.2
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-docx==1.2.0
pytz==2025.2
six==1.15.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
whitenoise==6.7.0
